requirements.txt
matplotlib~=3.10.3
python-dateutil
rich~=14.0.0
fpdf2